refactor(renderer): separate calendar markup from data control [skip ci]
